Big batch of changes
Noweb tangling is enabled, that should make it easier to embed scheme
code inside of elisp and the other way around.  see run-e4l for an
example.

Added e4l-add-server-path to properly add the paths to the osc server
on start-time rather than init time.

Scheme side prop listener switched to a lambda style when asking for
properties, which is think is the right idea there.

The live object browser is getting a lot closer to being fully
interactive:
- added a e4l-object-browse-at-point function, which will browse the
thing at the point if it knows about it.
- Properties are queried and displayed
- Childs get their id's displayed and you can browse them
- Children get a list of their ids displayed